youtube_mcp/requirements.txt
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
fastmcp>=0.1.0
pydantic>=2.0.0